+++ tests/rowQuantiles_subset.R
@@ -13,17 +13,16 @@ rowQuantiles_R <- function(x, probs, na.rm = FALSE, drop = TRUE, ..., useNames =
       rep(na_value, times = length(probs))
 
     } else {
-      as.vector(quantile(x, probs = probs, na.rm = na.rm, ...))
+      as.vector(quantile(x, probs = probs, na.rm = na.rm, names = FALSE, ...))
     }
   }, probs = probs, na.rm = na.rm)
 
   if (!is.null(dim(q))) q <- t(q)
   else dim(q) <- c(nrow(x), length(probs))
 
-  digits <- max(2L, getOption("digits"))
-  colnames(q) <- sprintf("%.*g%%", digits, 100 * probs)
+  colnames(q) <- matrixStats:::quantile_probs_names(probs)
   rownames(q) <- rownames(x)
-  if (isFALSE(useNames)) rownames(q) <- NULL
+  if (isFALSE(useNames)) dimnames(q) <- NULL
 
   if (drop) q <- drop(q)
   q